Crate fire

Implementation of the Fast-Inertial-Relaxation-Engine (FIRE) algorithm

References

Bitzek, E. et al. Structural Relaxation Made Simple. Phys. Rev. Lett. 2006, 97 (17), 170201.

http://users.jyu.fi/~pekkosk/resources/pdf/FIRE.pdf

https://github.com/siesta-project/flos/blob/master/flos/optima/fire.lua

Modules

line
Line search, also called one-dimensional search, refers to an optimization procedure for univariable functions.

lj

Structs

CachedProblem
Represents an optimization problem with cache for avoiding unnecessary function re-evaluations.

FIRE
The Fast-Inertial-Relaxation-Engine (FIRE) algorithm

Progress
Important iteration data in minimization, useful for progress monitor or defining artificial termination criteria.

UserTermination
For user defined termination criteria

Traits

GradientBasedMinimizer
Common interfaces for structure relaxation

TerminationCriteria
Define termination conditions

Functions

fire
Create FIRE optimization interface

monitor
Create user defined monitor or termination criteria with a closure