Function fire::line::mcstep::update_trial_interval

pub(crate) fn update_trial_interval(
    x: &mut f64, 
    fx: &mut f64, 
    dx: &mut f64, 
    y: &mut f64, 
    fy: &mut f64, 
    dy: &mut f64, 
    t: &mut f64, 
    ft: f64, 
    dt: f64, 
    tmin: f64, 
    tmax: f64, 
    brackt: &mut bool
) -> Result<i32, Error>

Update a safeguarded trial value and interval for line search.

This function assumes that the derivative at the point of x in the direction of the step. If the bracket is set to true, the minimizer has been bracketed in an interval of uncertainty with endpoints between x and y.

Arguments

x, fx, and dx: variables which specify the step, the function, and the derivative at the best step obtained so far. The derivative must be negative in the direction of the step, that is, dx and t-x must have opposite signs. On output these parameters are updated appropriately.

y, fy, and dy: variables which specify the step, the function, and the derivative at the other endpoint of the interval of uncertainty. On output these parameters are updated appropriately.

t, ft, and dt: variables which specify the step, the function, and the derivative at the current step. If bracket is set true then on input t must be between x and y. On output t is set to the new step.

tmin, tmax: lower and upper bounds for the step.

brackt: Specifies if a minimizer has been bracketed. If the minimizer has not been bracketed then on input brackt must be set false. If the minimizer is bracketed then on output brackt is set true.

Return

Status value. Zero indicates a normal termination.