Trait fire::line::LineSearch

pub trait LineSearch<E> where
    E: FnMut(f64, &mut f64) -> f64, {
    fn find(&mut self, step: &mut f64, phi: E) -> Result<usize>;
}

Required methods

fn find(&mut self, step: &mut f64, phi: E) -> Result<usize>

Given initial step size and phi function, returns an satisfactory step size.

step is a positive scalar representing the step size along search direction. phi is an univariable function of step for evaluating the value and the gradient projected onto search direction.

Loading content...