Struct fire::CachedProblem

pub struct CachedProblem<E> where
    E: FnMut(&[f64], &mut [f64]) -> f64,  { /* fields omitted */ }

Represents an optimization problem with cache for avoiding unnecessary function re-evaluations.

Examples

let mut x = vec![0.0; 5];
let mut prb = CachedProblem::new(&x, f);
let d = [0.2; 5];
prb.take_line_step(&d, 1.0);
let fx = prb.value();
let gx = prb.gradient();

Methods

impl<E> CachedProblem<E> where
E: FnMut(&[f64], &mut [f64]) -> f64,

pub fn new(x: &[f64], f: E) -> Self

Construct a CachedProblem

Parameters

x: initial position

f: a closure for function evaluation of value and gradient.

pub fn ncalls(&self) -> usize

The number of function calls

pub fn take_line_step(&mut self, displ: &[f64], step: f64)

Update position x at a prescribed displacement and step size.

x += step * displ

pub fn value(&mut self) -> f64

Return function value at current position.

The function will be evaluated when necessary.

pub fn value_prev(&self) -> f64

Return function value at previous point

pub fn gradient_prev(&self) -> &[f64]

pub fn gradient(&mut self) -> &[f64]

Return a reference to gradient at current position.

The function will be evaluated when necessary.

pub fn position(&self) -> &[f64]

Return a reference to current position vector.

pub fn set_position(&mut self, p: &[f64])

Update position from p

pub fn revert(&mut self)

Revert to previous point