Struct fire::FIRE

pub struct FIRE {
    pub max_step: f64,
    pub f_alpha: f64,
    pub alpha_start: f64,
    pub f_inc: f64,
    pub f_dec: f64,
    pub n_min: usize,
    // some fields omitted
}

The Fast-Inertial-Relaxation-Engine (FIRE) algorithm

Notes from the paper

n_min larger than 1 (at least a few smooth steps after freezing);

f_inc larger than but near to one (avoid too fast acceleration);

f_dec smaller than 1 but much larger than zero (avoid too heavy slowing down)

alpha_start larger than, but near to zero (avoid too heavy damping)

f_alpha smaller than, but near to one (mixing is efficient some time after restart).

Fields
max_step: f64

The maximum size for an optimization step. According to the paper, this is the only parameter needs to be adjusted by the user.

The default value is 0.10.

f_alpha: f64

Factor used to decrease alpha-parameter if downhill

The default value is 0.99.

alpha_start: f64

Initial alpha-parameter.

The default value is 0.1.

f_inc: f64

Factor used to increase time-step if downhill

The default value is 1.1.

f_dec: f64

Factor used to decrease time-step if uphill

The default value is 0.5.

n_min: usize

Minimum number of iterations ("latency" time) performed before time-step may be increased, which is important for the stability of the algorithm.

The default value is 5.

Methods

impl FIRE

pub fn with_max_step(self, maxstep: f64) -> Self

Set the maximum size for an optimization step.

pub fn with_md_scheme(self, scheme: &str) -> Self

Set MD scheme for position and velocity update

pub fn with_max_cycles(self, n: usize) -> Self

Set the maximum cycles for termination.

pub fn with_max_gradient_norm(self, gmax: f64) -> Self

Set the maximum gradient/force norm for termination.

pub fn with_line_search(self) -> Self

Enable line search of optimal step size.

The default is no line search.